+++ lib/domain/entities/product/product_mini.dart
@@ -11,7 +11,7 @@ class ProductMini {
   final int id;
   final String title;
   final double ourRating;
-  final double discount;
+  final int discount;
   final double normalPrice;
   final String imageMini;
 }
